0.16.2 phase B: per-namespace dispatch carve-out
Phase A lifted the helper domains (factories, http, lm, replay_glue)
out of `imports.rs`, but the 29-arm `dispatch_aver_import` match
still lived inline in the 870-line file. This commit carves it
into eight per-namespace submodules — each effect family lives
next to its decoders / factories / glue, and the top-level dispatch
becomes a chained handoff that stays readable as new namespaces land.

Layout:

  imports.rs                  101 lines  (dispatch chain + re-exports)
  imports/args.rs              60        (args_len, args_get)
  imports/console.rs          113        (print, error, warn, readLine)
  imports/disk.rs             191        (read/write/append text, exists,
                                          delete, deleteDir, listDir, makeDir)
  imports/env.rs               52        (env_get, env_set)
  imports/numeric.rs          102        (Random.int / .float + float math)
  imports/tcp.rs              218        (connect, writeLine, readLine,
                                          close, send, ping)
  imports/terminal.rs         206        (12 terminal_* arms)
  imports/time.rs              75        (now, unixMs, sleep)

  imports/factories.rs        422        (host_*_make builders)
  imports/http.rs             153        (HttpVerb + simple/body dispatch)
  imports/lm.rs               156        (LM transport + host_print)
  imports/replay_glue.rs      129        (try_replay / record / json_*)

Pattern per namespace: `pub(super) fn dispatch(name, caller, params,
results) -> Result<bool>` walks its own match arms, returns `true`
on hit and `false` on miss. `dispatch_aver_import` chains them in
fixed order; HTTP verbs stay inline as one-line aliases.

Visibility upshift: decoders in `decode.rs` moved from `pub(super)`
to `pub(crate)` so per-namespace files can reach them via
`super::super::decode::*`. Same change already applied to the
helpers in phase A.

8/8 smokes pass, both clippy invocations clean. No behavioural
change — every effect lands in the same `aver_rt::*` call as
before; the dispatch chain just walks a few extra `if`s before
finding its match arm (one per namespace, dropped early).
